+++ src/shared/Router.jsx
@@ -1,16 +1,13 @@
 import React from "react";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 
-import Login from "../login/Login";
-
-import Signup from "../login/Signup";
+import Login from "../login/LoginPage";
 
 const Router = () => {
   return (
     <BrowserRouter>
       <Routes>
         <Route path="/login" element={<Login />} />
-        <Route path="/signup" element={<Signup />} />
       </Routes>
     </BrowserRouter>
   );
